+++ Reach/src/utopia/reach/container/multi/NonContextualViewContainerFactory.scala
@@ -3,7 +3,7 @@ package utopia.reach.container.multi
 import utopia.flow.view.template.eventful.Changing
 import utopia.reach.component.factory.ComponentFactoryFactory.Cff
 import utopia.reach.component.template.ReachComponent
-import utopia.reach.component.wrapper.ComponentCreationResult.SwitchableCreations
+import utopia.reach.component.wrapper.ComponentCreationResult.SwitchableComponents
 import utopia.reach.component.wrapper.ComponentWrapResult.SwitchableComponentsWrapResult
 import utopia.reach.component.wrapper.Open
 
@@ -30,8 +30,8 @@ trait NonContextualViewContainerFactory[+Container <: ReachComponent, -Top <: Re
 	  * @return The created container, created components and the additional result
 	  */
 	def build[F, C <: Top, R](contentFactory: Cff[F])
-	                         (fill: Iterator[F] => SwitchableCreations[C, R]): SwitchableComponentsWrapResult[Container, C, R] =
-		apply(Open.manyUsing[F, C, Changing[Boolean], R](contentFactory)(fill)(hierarchy.top))
+	                         (fill: Iterator[F] => SwitchableComponents[C, R]): SwitchableComponentsWrapResult[Container, C, R] =
+		apply(Open.manyConditionalUsing[F, C, R](contentFactory)(fill)(hierarchy.top))
 	
 	/**
 	  * Builds a new container which reflects the contents of a multi-value pointer.